Move code to separate files
